requirements.txt
google-generativeai
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
python-dotenv
asyncpg
passlib[bcrypt]
python-jose[cryptography]
pandas
email-validator
